+++ src/main/java/com/toy4/domain/dutyHistory/dto/request/DutyCancellationRequest.java
@@ -1,5 +1,6 @@
-package com.toy4.domain.dayOffHistory.dto;
+package com.toy4.domain.dutyHistory.dto.request;
 
+import com.toy4.domain.dutyHistory.dto.DutyHistoryMainDto;
 import com.toy4.domain.schedule.RequestStatus;
 import lombok.Getter;
 
@@ -7,15 +8,15 @@ import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
 @Getter
-public class DayOffCancellationRequest {
+public class DutyCancellationRequest {
 
     @NotNull
     private Long employeeId;
     @NotBlank
     private String status;
 
-    public DayOffHistoryMainDto toDto() {
-        return DayOffHistoryMainDto.builder()
+    public DutyHistoryMainDto toDto() {
+        return DutyHistoryMainDto.builder()
                 .employeeId(employeeId)
                 .status(RequestStatus.getByDescription(status))
                 .build();
