+++ src/main/java/com/toy4/domain/dutyHistory/dto/request/DutyRegistrationRequest.java
@@ -1,23 +1,27 @@
-package com.toy4.domain.dutyHistory.dto;
+package com.toy4.domain.dutyHistory.dto.request;
 
-import com.toy4.domain.schedule.RequestStatus;
+import com.toy4.domain.dutyHistory.dto.DutyHistoryMainDto;
 import lombok.Getter;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
+import java.time.LocalDate;
 
 @Getter
-public class DutyCancellationRequest {
+public class DutyRegistrationRequest {
 
     @NotNull
     private Long employeeId;
+    @NotNull
+    private LocalDate date;
     @NotBlank
-    private String status;
+    private String reason;
 
     public DutyHistoryMainDto toDto() {
         return DutyHistoryMainDto.builder()
                 .employeeId(employeeId)
-                .status(RequestStatus.getByDescription(status))
+                .date(date)
+                .reason(reason)
                 .build();
     }
 }
