+++ src/main/java/com/toy4/domain/schedule/ScheduleMainService.java
@@ -1,8 +1,8 @@
 package com.toy4.domain.schedule;
 
-import com.toy4.domain.dayOffHistory.dto.FindDayOffHistoryResponse;
+import com.toy4.domain.dayOffHistory.dto.response.FindDayOffHistoryResponse;
 import com.toy4.domain.dayOffHistory.repository.DayOffHistoryRepository;
-import com.toy4.domain.dutyHistory.dto.FindDutyHistoryResponse;
+import com.toy4.domain.dutyHistory.dto.response.FindDutyHistoryResponse;
 import com.toy4.domain.dutyHistory.repository.DutyHistoryRepository;
 import com.toy4.domain.employee.domain.Employee;
 import com.toy4.domain.employee.exception.EmployeeException;
