+++ src/main/java/com/toy4/domain/schedule/ScheduleResponse.java
@@ -1,7 +1,7 @@
 package com.toy4.domain.schedule;
 
-import com.toy4.domain.dayOffHistory.dto.FindDayOffHistoryResponse;
-import com.toy4.domain.dutyHistory.dto.FindDutyHistoryResponse;
+import com.toy4.domain.dayOffHistory.dto.response.FindDayOffHistoryResponse;
+import com.toy4.domain.dutyHistory.dto.response.FindDutyHistoryResponse;
 import lombok.Builder;
 import lombok.Getter;
